Move files to make it cleaner
